+++ tinker_cookbook/recipes/search_tool/search_env.py
@@ -1,21 +1,36 @@
-"""Search tool utilities.
+"""Search environment and task utilities.
 
-This module provides utility functions and data types for search-based RL tasks.
-Includes dataset loading, answer normalization, and task instructions.
-
-Note: SearchEnv, SearchR1Dataset, and SearchR1DatasetBuilder have been replaced
-by the new tool-use implementation in train.py using AgentToolMessageEnv.
+This module provides:
+- Task definition: SEARCH_TASK_INSTRUCTIONS, SearchR1Datum, dataset loading
+- Environment construction: build_env(), EnvGroupBuilder, RLDataset, RLDatasetBuilder
 """
 
+from __future__ import annotations
+
 import os
+import random
 import re
 import string
+from dataclasses import dataclass
 from functools import reduce
 from pathlib import Path
-from typing import Literal, TypedDict, cast
+from typing import Literal, Sequence, TypedDict, cast
 
 import pandas as pd
 from huggingface_hub import hf_hub_download
+
+from tinker_cookbook import model_info, tokenizer_utils
+from tinker_cookbook.recipes.search_tool.tools import (
+    ChromaTool,
+    ChromaToolConfig,
+    TextAnswerReward,
+    build_tools,
+)
+from tinker_cookbook.renderers import get_renderer
+from tinker_cookbook.renderers.base import Message, Renderer
+from tinker_cookbook.rl import types
+from tinker_cookbook.rl.message_env import EnvFromMessageEnv
+from tinker_cookbook.tool_use import AgentToolMessageEnv
 
 SEARCH_TASK_INSTRUCTIONS = """You are an expert assistant who solves tasks using a Wikipedia search tool.
 
@@ -132,3 +147,150 @@ def download_search_r1_dataset(split: Literal["train", "test"]) -> list[SearchR1
     df_raw = pd.read_parquet(local_parquet_filepath)
 
     return df_raw.apply(process_single_row, axis=1).tolist()
+
+
+
+def _initial_messages(
+    datum: SearchR1Datum,
+    renderer: Renderer,
+    chroma_tool: ChromaTool,
+) -> list[Message]:
+    """Build initial messages with tool schemas and task question."""
+    # Get tool schema from our @tool-decorated method
+    tool_schemas = [chroma_tool.search.to_spec()]
+    prefix = renderer.create_conversation_prefix_with_tools(
+        tools=tool_schemas,
+        system_prompt=SEARCH_TASK_INSTRUCTIONS,
+    )
+    return prefix + [{"role": "user", "content": datum["question"]}]
+
+
+def build_env(
+    datum: SearchR1Datum,
+    model_name: str,
+    *,
+    renderer_name: str | None = None,
+    max_turns: int = 5,
+    chroma_tool: ChromaTool,
+    format_coef: float = 0.1,
+) -> EnvFromMessageEnv:
+    """Build an RL environment for a single search task."""
+    tokenizer = tokenizer_utils.get_tokenizer(model_name)
+    chosen_renderer = renderer_name or model_info.get_recommended_renderer_name(model_name)
+    renderer = get_renderer(chosen_renderer, tokenizer)
+
+    tools = build_tools(chroma_tool)
+    msg_env = AgentToolMessageEnv(
+        tools=tools,
+        initial_messages=_initial_messages(datum, renderer, chroma_tool),
+        max_turns=max_turns,
+        reward_fn=TextAnswerReward(gold_answers=datum["answer"], format_coef=format_coef),
+    )
+    return EnvFromMessageEnv(
+        renderer=renderer,
+        message_env=msg_env,
+        failed_parse_reward=-0.1,
+    )
+
+
+class EnvGroupBuilder(types.EnvGroupBuilder):
+    """EnvGroupBuilder that creates search environments with a shared ChromaTool."""
+
+    def __init__(
+        self,
+        datum: SearchR1Datum,
+        model_name: str,
+        renderer_name: str | None,
+        max_turns: int,
+        group_size: int,
+        chroma_tool: ChromaTool,
+        format_coef: float = 0.1,
+    ):
+        self.datum = datum
+        self.model_name = model_name
+        self.renderer_name = renderer_name
+        self.max_turns = max_turns
+        self.group_size = group_size
+        self.chroma_tool = chroma_tool
+        self.format_coef = format_coef
+
+    async def make_envs(self) -> Sequence[types.Env]:
+        return [
+            build_env(
+                datum=self.datum,
+                model_name=self.model_name,
+                renderer_name=self.renderer_name,
+                max_turns=self.max_turns,
+                chroma_tool=self.chroma_tool,
+                format_coef=self.format_coef,
+            )
+            for _ in range(self.group_size)
+        ]
+
+    def logging_tags(self) -> list[str]:
+        return [self.datum.get("data_source", "unknown")]
+
+
+class RLDataset(types.RLDataset):
+    """Dataset that processes search EnvGroupBuilders once per epoch."""
+
+    def __init__(
+        self,
+        env_group_builders: list[EnvGroupBuilder],
+        batch_size: int,
+    ):
+        self.env_group_builders = env_group_builders
+        self.batch_size = batch_size
+
+    def get_batch(self, index: int) -> Sequence[types.EnvGroupBuilder]:
+        start = index * self.batch_size
+        end = start + self.batch_size
+        return self.env_group_builders[start:end]
+
+    def __len__(self) -> int:
+        return len(self.env_group_builders) // self.batch_size
+
+
+@dataclass(frozen=True)
+class RLDatasetBuilder(types.RLDatasetBuilder):
+    """Build an RL dataset over SearchR1 tasks with ChromaTool."""
+
+    model_name: str
+    chroma_config: ChromaToolConfig
+    split: Literal["train", "test"] = "train"
+    max_tasks: int | None = 100
+    batch_size: int = 4
+    group_size: int = 4
+    renderer_name: str | None = None
+    max_turns: int = 5
+    format_coef: float = 0.1
+    seed: int = 0
+
+    async def __call__(self) -> tuple[types.RLDataset, types.RLDataset | None]:
+        # Create shared ChromaTool
+        chroma_tool = await ChromaTool.create(self.chroma_config)
+
+        # Load and shuffle tasks
+        data = download_search_r1_dataset(self.split)
+        rng = random.Random(self.seed)
+        rng.shuffle(data)
+        if self.max_tasks is not None:
+            data = data[: self.max_tasks]
+
+        env_builders = [
+            EnvGroupBuilder(
+                datum=datum,
+                model_name=self.model_name,
+                renderer_name=self.renderer_name,
+                max_turns=self.max_turns,
+                group_size=self.group_size,
+                chroma_tool=chroma_tool,
+                format_coef=self.format_coef,
+            )
+            for datum in data
+        ]
+        dataset = RLDataset(
+            env_group_builders=env_builders,
+            batch_size=self.batch_size,
+        )
+        return dataset, None
